refactor(canvas): move conversation stores to runtime
