+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatSessionsRail.tsx
@@ -8,7 +8,7 @@ import { Button, TextField } from '../ui';
 import { useI18n } from '../../i18n';
 import { ChatSessionRailItem } from './ChatSessionRailItem';
 import type { WorkspaceOption } from './types';
-import type { ConversationCompletionStatus } from './hooks/conversationCompletionStore';
+import type { ConversationCompletionStatus } from '../../agent-chat/runtime/conversationCompletionStore';
 
 const SESSION_PREVIEW_LIMIT = 10;
 const GLOBAL_CHAT_ID = '__global_chat__';
