+++ apps/canvas-workspace/src/renderer/src/components/chat/hooks/useScopeRunningSessions.test.tsx
@@ -8,7 +8,7 @@ import {
   resetConversationStoreForTests,
   setConversationLoading,
   setConversationMessages,
-} from './conversationStore';
+} from '../../../agent-chat/runtime/conversationStore';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
