+++ apps/canvas-workspace/src/renderer/src/components/chat/hooks/useStableSessionRail.ts
@@ -5,7 +5,7 @@ import { useI18n } from '../../../i18n';
 import type { AgentSessionInfo } from '../../../types';
 import type { UnifiedSession } from '../ChatSessionsRail';
 import type { AgentScope, OtherWorkspaceSession, WorkspaceOption } from '../types';
-import type { ConversationCompletionStatus } from './conversationCompletionStore';
+import type { ConversationCompletionStatus } from '../../../agent-chat/runtime/conversationCompletionStore';
 
 interface UseStableSessionRailOptions {
   agentScope: AgentScope;
