+++ apps/canvas-workspace/src/renderer/src/components/shell/ConversationCompletionToastBridge/index.tsx
@@ -6,7 +6,7 @@ import {
   isConversationVisible,
   markConversationCompletionNotified,
   useConversationCompletions,
-} from '../../chat/hooks/conversationCompletionStore';
+} from '../../../agent-chat/runtime/conversationCompletionStore';
 
 /** Always-mounted, low-interruption feedback for genuinely background turns. */
 export const ConversationCompletionToastBridge = () => {
